// src/app/model/cartForm.ts
export class CartForm {
  CartId: string ;
  RegisterDate?: Date ;
  LastName: string ;
  FirstName: string ;
  IdentityNumber: string ;
  BirthDate?: Date;
  GenderId?: number;
  Address: string ;
  CityName: string;
  CellPhone: string ;
  AllowToSendSMS: boolean;
  Email: string ;
  AllowToSendEmail: boolean ;
  Phone: string ;
  Phone2: string ;
  IsHistadrutMember: boolean;

};